+++ videoanalyst/config/config.py
@@ -24,8 +24,10 @@ for task in task_list:
     if task == "track":
         test_cfg[task]["model"] = get_model_cfg(task_list)[task]
     elif task == "vos":
-        test_cfg[task]["tracker"] = get_model_cfg(task_list)['vos']
         test_cfg[task]["segmenter"] = get_model_cfg(task_list)['vos']
+        test_cfg[task]["tracker_model"] = get_model_cfg(task_list)['track']
+        test_cfg[task]["tracker_pipeline"] = get_pipeline_cfg(
+            task_list)['track']
 
     test_cfg[task]["pipeline"] = get_pipeline_cfg(task_list)[task]
     test_cfg[task]["tester"] = get_tester_cfg(task_list)[task]
@@ -40,7 +42,11 @@ for task in task_list:
     train_cfg[task]["exp_save"] = default_str
     train_cfg[task]["num_processes"] = 1  #number of devices
     train_cfg[task]["device"] = "cuda"  #[cuda|cpu]
-    train_cfg[task]["model"] = get_model_cfg(task_list)[task]
+    if task == "track":
+        train_cfg[task]["model"] = get_model_cfg(task_list)[task]
+    elif task == "vos":
+        train_cfg[task]["tracker_model"] = get_model_cfg(task_list)[task]
+        train_cfg[task]["segmenter"] = get_model_cfg(task_list)[task]
     train_cfg[task]["pipeline"] = get_pipeline_cfg(task_list)[task]
     train_cfg[task]["tester"] = get_tester_cfg(task_list)[task]
     train_cfg[task]["data"] = get_data_cfg(task_list)[task]