+++ videoanalyst/data/transformer/transformer_impl/random_crop_transformer.py
@@ -2,6 +2,7 @@ from typing import Dict
 
 import cv2
 import numpy as np
+import torch
 from yacs.config import CfgNode
 
 from videoanalyst.data.utils.crop_track_pair import crop_track_pair
@@ -18,13 +19,27 @@ class RandomCropTransformer(TransformerBase):
     Hyper-parameters
     ----------------
 
+    context_amount: float
+        the context factor for template image
+    max_scale: float
+        the max scale change ratio for search image
+    max_shift:  float
+        the max shift change ratio for search image
+    max_scale_temp: float
+        the max scale change ratio for template image
+    max_shift_temp:  float
+        the max shift change ratio for template image
+    z_size: int
+        output size of template image
+    x_size: int
+        output size of search image
     """
     default_hyper_params = dict(
         context_amount=0.5,
         max_scale=0.3,
         max_shift=0.4,
-        max_scale_temp=0,
-        max_shift_temp=0,
+        max_scale_temp=0.0,
+        max_shift_temp=0.0,
         z_size=127,
         x_size=303,
     )
@@ -42,12 +57,13 @@ class RandomCropTransformer(TransformerBase):
         data2 = sampled_data["data2"]
         im_temp, bbox_temp = data1["image"], data1["anno"]
         im_curr, bbox_curr = data2["image"], data2["anno"]
-        im_z, bbox_z, im_x, bbox_x = crop_track_pair(im_temp,
-                                                     bbox_temp,
-                                                     im_curr,
-                                                     bbox_curr,
-                                                     config=self._hyper_params,
-                                                     rng=self._state["rng"])
+        im_z, bbox_z, im_x, bbox_x, _, _ = crop_track_pair(
+            im_temp,
+            bbox_temp,
+            im_curr,
+            bbox_curr,
+            config=self._hyper_params,
+            rng=self._state["rng"])
 
         sampled_data["data1"] = dict(image=im_z, anno=bbox_z)
         sampled_data["data2"] = dict(image=im_x, anno=bbox_x)