+++ videoanalyst/engine/trainer/builder.py
@@ -15,7 +15,8 @@ from ..monitor.monitor_base import TASK_MONITORS
 from .trainer_base import TASK_TRAINERS, TrainerBase
 
 
-def build(task: str, cfg: CfgNode, optimizer, dataloader) -> TrainerBase:
+def build(task: str, cfg: CfgNode, optimizer, dataloader,
+          tracker=None) -> TrainerBase:
     r"""
     Builder function.
 
@@ -42,7 +43,10 @@ def build(task: str, cfg: CfgNode, optimizer, dataloader) -> TrainerBase:
         monitors = []
 
     name = cfg.name
-    trainer = MODULES[name](optimizer, dataloader, monitors)
+    if task == "vos":
+        trainer = MODULES[name](optimizer, dataloader, monitors, tracker)
+    else:
+        trainer = MODULES[name](optimizer, dataloader, monitors)
     hps = trainer.get_hps()
     hps = merge_cfg_into_hps(cfg[name], hps)
     trainer.set_hps(hps)