+++ videoanalyst/engine/trainer/trainer_impl/regular_trainer.py
@@ -61,15 +61,6 @@ class RegularTrainer(TrainerBase):
         self._state["initialized"] = False
         self._state["devices"] = torch.device("cuda:0")
 
-    def update_params(self, ):
-        super(RegularTrainer, self).update_params()
-        self._hyper_params["num_iterations"] = self._hyper_params[
-            "nr_image_per_epoch"] // self._hyper_params["minibatch"]
-        self._state["snapshot_dir"] = osp.join(self._hyper_params["exp_save"],
-                                               self._hyper_params["exp_name"])
-
-        self._state["snapshot_file"] = self._hyper_params["snapshot"]
-
     def init_train(self, ):
         torch.cuda.empty_cache()
         # move model & loss to target devices