+++ videoanalyst/evaluation/utils/misc.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*
-
-from yacs.config import CfgNode as CN
-
-from videoanalyst.model import builder as model_builder
-from videoanalyst.pipeline import builder as pipeline_builder
-from videoanalyst.utils.misc import load_cfg